# service/api_pods_podid.py
from fastapi import APIRouter
from models import Pod, NewPod, UpdatePod, PodResponse, Password, DeletePodResponse
from channels import CommandChannel
from tapisservice.tapisfastapi.utils import g, ok

from tapisservice.logs import get_logger
logger = get_logger(__name__)

router = APIRouter()


#### /pods/{pod_id}

@router.put("/pods/{pod_id}", tags=["Pods"], summary="update_pod", operation_id="update_pod", response_model=PodResponse)
async def update_pod(pod_id, update_pod: UpdatePod):
    logger.info(f"UPDATE /pods/{pod_id} - Top of update_pod.")

    pod = Pod.db_get_DAO(pod_id)

    # Do checks here, ensure pod exists. etc.

    pod = Pod(**update_pod.dict())

    # Do more update things.

    return ok("update_pod - Not yet implemented.")


@router.delete("/pods/{pod_id}", tags=["Pods"], summary="delete_pod", operation_id="delete_pod", response_model=DeletePodResponse)
async def delete_pod(pod_id):
    logger.info(f"DELETE /pods/{pod_id} - Top of delete_pod.")

    # Needs to delete pod, service, db_pod, db_password
    pod = Pod.db_get_with_pk(pod_id, tenant=g.request_tenant_id, site=g.site_id)
    password = Password.db_get_with_pk(pod_id, tenant=g.request_tenant_id, site=g.site_id)

    pod.db_delete()
    password.db_delete()

    return ok(result=None, msg="Pod successfully deleted.")


@router.get("/pods/{pod_id}", tags=["Pods"], summary="get_pod", operation_id="get_pod", response_model=PodResponse)
async def get_pod(pod_id):
    logger.info(f"GET /pods/{pod_id} - Top of get_pod.")

    # TODO .display(), search, permissions

    pod = Pod.db_get_with_pk(pod_id, tenant=g.request_tenant_id, site=g.site_id)

    return ok(result=pod.display(), msg="Pod retrieved successfully.")